move page js to the page dir
